Internals

Initialization

Missing docstring.

Missing docstring for IJulia.init. Check Documenter's build log for details.

Cell execution hooks

IJulia.pop_posterror_hook — Function

pop_posterror_hook(f::Function)

Remove a function f() from the list of functions to execute after an error occurs when a notebook cell is evaluated.

IJulia.pop_postexecute_hook — Function

pop_postexecute_hook(f::Function)

Remove a function f() from the list of functions to execute after executing any notebook cell.

IJulia.pop_preexecute_hook — Function

pop_preexecute_hook(f::Function)

Remove a function f() from the list of functions to execute before executing any notebook cell.

IJulia.push_posterror_hook — Function

pop_posterror_hook(f::Function)

Remove a function f() from the list of functions to execute after an error occurs when a notebook cell is evaluated.

IJulia.push_postexecute_hook — Function

push_postexecute_hook(f::Function)

Push a function f() onto the end of a list of functions to execute after executing any notebook cell.

IJulia.push_preexecute_hook — Function

push_preexecute_hook(f::Function)

Push a function f() onto the end of a list of functions to execute before executing any notebook cell.

Messaging

Missing docstring.

Missing docstring for IJulia.Msg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.msg_header. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.send_ipython. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.recv_ipython. Check Documenter's build log for details.

IJulia.set_cur_msg — Function

Jupyter associates cells with message headers. Once a cell's execution state has been set as to idle, it will silently drop stream messages (i.e. output to stdout and stderr) - see https://github.com/jupyter/notebook/issues/518. When using Interact, and a widget's state changes, a new message header is sent to the IJulia kernel, and while Reactive is updating Signal graph state, it's execution state is busy, meaning Jupyter will not drop stream messages if Interact can set the header message under which the stream messages will be sent. Hence the need for this function.

Missing docstring.

Missing docstring for IJulia.send_status. Check Documenter's build log for details.

Request handlers

Missing docstring.

Missing docstring for IJulia.handlers. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.connect_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.execute_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.shutdown_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.interrupt_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.inspect_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.history_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.complete_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.kernel_info_request. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.is_complete_request. Check Documenter's build log for details.

Event loop

Missing docstring.

Missing docstring for IJulia.eventloop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.waitloop. Check Documenter's build log for details.

IO

Missing docstring.

Missing docstring for IJulia.IJuliaStdio. Check Documenter's build log for details.

IJulia.capture_stdout — Constant

The IJulia kernel captures all stdout and stderr output and redirects it to the notebook. When debugging IJulia problems, however, it can be more convenient to not capture stdout and stderr output (since the notebook may not be functioning). This can be done by editing IJulia.jl to set capture_stderr and/or capture_stdout to false.

Missing docstring.

Missing docstring for IJulia.capture_stderr. Check Documenter's build log for details.

IJulia.watch_stream — Function

Continually read from (size limited) Libuv/OS buffer into an IObuffer to avoid problems when the Libuv/OS buffer gets full (https://github.com/JuliaLang/julia/issues/8789). Send data immediately when buffer contains more than max_bytes bytes. Otherwise, if data is available it will be sent every stream_interval seconds (see the Timers set up in watchstdio). Truncate the output to `maxoutputperrequest` bytes per execution request since excessive output can bring browsers to a grinding halt.

Multimedia display

Missing docstring.

Missing docstring for IJulia.InlineDisplay. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.InlineIOContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.ipy_mime. Check Documenter's build log for details.

IJulia.ijulia_mime_types — Constant

A vector of MIME types (or vectors of MIME types) that IJulia will try to render. IJulia will try to render every MIME type specified in the first level of the vector. If a vector of MIME types is specified, IJulia will include only the first MIME type that is renderable (this allows for the expression of priority and exclusion of redundant data).

For example, since "text/plain" is specified as a first-child of the array, IJulia will always try to include a "text/plain" representation of anything that is displayed. Since markdown and html are specified within a sub-vector, IJulia will always try to render "text/markdown", and will only try to render "text/html" if markdown isn't possible.

IJulia.ijulia_jsonmime_types — Constant

MIME types that when rendered (via stringmime) return JSON data. See ijulia_mime_types for a description of how MIME types are selected.

This is necessary to embed the JSON as is in the displaydata bundle (rather than as stringify'd JSON).

Missing docstring.

Missing docstring for IJulia.limitstringmime. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.israwtext. Check Documenter's build log for details.

IJulia.display_dict — Function

Generate a dictionary of mime_type => data pairs for all registered MIME types. This is the format that Jupyter expects in displaydata and executeresult messages.

IJulia.display_mimejson — Function

Generate the preferred json-MIME representation of x.

Returns a tuple with the selected MIME type and the representation of the data using that MIME type (as a JSONText).

IJulia.display_mimestring — Function

Generate the preferred MIME representation of x.

Returns a tuple with the selected MIME type and the representation of the data using that MIME type.

Missing docstring.

Missing docstring for IJulia.register_mime. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.register_jsonmime. Check Documenter's build log for details.

Jupyter

Missing docstring.

Missing docstring for IJulia.find_jupyter_subcommand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IJulia.launch. Check Documenter's build log for details.

Debugging

IJulia.set_verbose — Function

set_verbose(v=true)

This function enables (or disables, for set_verbose(false)) verbose output from the IJulia kernel, when called within a running notebook. This consists of log messages printed to the terminal window where jupyter was launched, displaying information about every message sent or received by the kernel. Used for debugging IJulia.

Utility

IJulia.num_utf8_trailing — Function

If d ends with an incomplete UTF8-encoded character, return the number of trailing incomplete bytes. Otherwise, return 0.